#274: comic components hooks
