+++ src/webui/components/comic/Editor.tsx
@@ -1,52 +1,21 @@
 
-import { useEffect, useRef, useState } from 'react';
+import { useRef } from 'react';
 
 import { Button, Column, Panel, Row } from '^/webui/designsystem';
-import { Editor } from '^/webui/editor';
+
+import useCreateHandler, { CancelHandler, CreateHandler } from './hooks/useCreateHandler';
+import useEditor from './hooks/useEditor';
 
 type Props = {
-    readonly onCreate: (imageData: string) => Promise<void>;
-    readonly onCancel?: () => void;
+    readonly onCreate: CreateHandler;
+    readonly onCancel?: CancelHandler;
 };
 
 export default function Component({ onCreate, onCancel }: Props)
 {
     const canvasRef = useRef<HTMLCanvasElement>(null);
-    const [editor, setEditor] = useState<Editor | undefined>(undefined);
-    const [creating, setCreating] = useState(false);
-
-    const handleCreate = async () =>
-    {
-        if (editor === undefined)
-        {
-            throw new Error('Editor not found');
-        }
-
-        const imageData = editor.export();
-
-        setCreating(true);
-
-        await onCreate(imageData);
-
-        setCreating(false);
-    };
-
-    useEffect(() =>
-    {
-        const canvas = canvasRef.current;
-
-        if (canvas === null)
-        {
-            throw new Error('Editor canvas not found');
-        }
-
-        const editor = new Editor(canvas);
-        editor.start();
-
-        setEditor(editor);
-
-        return () => { editor.stop(); };
-    }, []);
+    const editor = useEditor(canvasRef);
+    const [creating, handleCreate, handleCancel] = useCreateHandler(editor, onCreate, onCancel);
 
     return <Panel>
         <Column alignX='stretch'>
@@ -54,7 +23,7 @@ export default function Component({ onCreate, onCancel }: Props)
             <Row alignX='right'>
                 {
                     onCancel !== undefined
-                        ? <Button type='secondary' text='Cancel' onClick={onCancel} />
+                        ? <Button type='secondary' text='Cancel' onClick={handleCancel} />
                         : null
                 }
                 <Button type={creating ? 'disabled' : 'primary'} text={creating ? 'Creating' : 'Create'} onClick={handleCreate} />
